Recoded features from the master branch
Added functionality to import a database, and restore a database.

The database scheme (and the amount of rows in each table) will now
display in the sidebar, and automatically update when any changes are
made.

Improved the execution process, now displays the result of the query,
changes are only saved if an actual modification was made to the
database.

Improved modularity of the application, questions will be an independant
addition.

Added a favicon and a generic title